% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AI_INDICATOR}
\alias{AI_INDICATOR}
\title{Aleutian Islands Indicator}
\format{
An object of class \code{list} of length 5.
}
\usage{
AI_INDICATOR
}
\description{
ESR Indicator data for the Aleutian Islands
}
\keyword{datasets}
