% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Capitalize the first letter of a character string}
\usage{
capitalize_first(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string with the first letter capitalized.
}
\description{
This function takes a character string as input and returns a new string where the first letter is capitalized.
}
