% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chapter_settings}
\alias{chapter_settings}
\title{Chapter Settings}
\format{
A nested list containing species groups included in each ESR chapters
 \describe{
   \item{misc_species}{List of miscellaneous species group names (e.g. Echinoderms, Shrimps, Eelpouts, Poachers).}
   \item{jellyfish}{List of jellyfish group names.}
   \item{structural_epifauna}{List of structural epifauna group names (e.g. Sponges, Sea anemones, Gorgonians, Pennatulaceans, Hydrocorals, Soft corals).}
   \item{forage_fish}{List of forage fish species group names (e.g. eulachon, capelin, pricklebacks, sandlance, Pacific sandfish)}
   }
}
\usage{
chapter_settings
}
\description{
Information about species groups included within different chapters.
}
\keyword{datasets}
