% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EBS_INDICATOR}
\alias{EBS_INDICATOR}
\title{Eastern Bering Sea Indicator}
\format{
An object of class \code{list} of length 4.
}
\usage{
EBS_INDICATOR
}
\description{
ESR Indicator data for the Eastern Bering Sea
}
\keyword{datasets}
