% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_figures.R
\name{extract_figures}
\alias{extract_figures}
\title{Extract figure info from ESR XML template}
\usage{
extract_figures(x)
}
\arguments{
\item{x}{A character vector containing XML data.}
}
\value{
A list containing information about figures, including figure number, path, and caption.
}
\description{
This function extracts information about figures from XML <figureNumber>, <figureBasename>, and <figureCaption> tags.
}
