% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tag.R
\name{extract_tag}
\alias{extract_tag}
\title{Extract content between ESR XML template tags}
\usage{
extract_tag(x, tag, remove_tabs = TRUE)
}
\arguments{
\item{x}{A character vector containing XML data.}

\item{tag}{The XML tag to extract content from (e.g. "indicatorTitle").}

\item{remove_tabs}{A logical value indicating whether to remove tabs from the extracted content (default is TRUE).}
}
\value{
A character vector containing the extracted content between the specified XML tags.
}
\description{
This function extracts the content between XML tags specified as a character vector.
}
