% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rema.R
\name{fit_rema_region}
\alias{fit_rema_region}
\title{Wrapper function to fit rema to region level biomass indices}
\usage{
fit_rema_region(x)
}
\arguments{
\item{x}{A data.frame containing SURVEY, YEAR, BIOMASS_MT, and CV columns}
}
\description{
Wrapper function to fit rema to region level biomass indices
}
