% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connected}
\alias{get_connected}
\title{Get Connected Oracle Channel}
\usage{
get_connected(channel = NULL, schema = NA)
}
\arguments{
\item{channel}{An open RODBC channel-- only used within functions to handle open connections.}

\item{schema}{A character string specifying the Oracle schema. If not provided, it prompts the user to enter the schema.}
}
\value{
An RODBC channel object representing the connection to the Oracle database.
}
\description{
Establishes a connection to an Oracle database channel using the specified schema, username, and password.
}
\details{
This function connects to an Oracle database using the provided schema, username, and password. If the schema is not provided,
it prompts the user to enter the Oracle schema. The function returns an RODBC channel object representing the connection.
}
