% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_codes.R
\name{get_group_codes}
\alias{get_group_codes}
\title{Get AFSC species codes for each taxonomic group}
\usage{
get_group_codes(group, rank, tax_table)
}
\arguments{
\item{group}{A character string designating the taxonomic group}

\item{rank}{A character string designating the taxonomic rank of grouping}

\item{tax_table}{Taxonomic classification table from Oracle database}
}
\value{
A numeric vector of species codes
}
\description{
This function takes a taxonomic grouping in the form of a character string and returns a vector of species codes that are found within that taxonomic group.
}
