% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOA_INDICATOR}
\alias{GOA_INDICATOR}
\title{Gulf of Alaska Indicator}
\format{
An object of class \code{list} of length 5.
}
\usage{
GOA_INDICATOR
}
\description{
ESR Indicator data for the Gulf of Alaska
}
\keyword{datasets}
