% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_tables.R
\name{make_index_tables}
\alias{make_index_tables}
\title{Make tables with ESR indicator time series}
\usage{
make_index_tables(x)
}
\arguments{
\item{x}{Indicator list (e.g. esrindex::AI_INDICATOR)}
}
\description{
This function writes abundance indicator results to .csv (./plots/[region]/[indicator]_indicator_table.csv)
}
