% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subarea_plot.R
\name{make_subarea_plot}
\alias{make_subarea_plot}
\title{Create Subarea-specific Biomass Plots}
\usage{
make_subarea_plot(
  indicator_data,
  indicator_name,
  bar_color = "#0085CA",
  error_bars = TRUE
)
}
\arguments{
\item{indicator_data}{A list containing indicator data, including timeseries and mean_sd.}

\item{indicator_name}{A character string specifying the indicator name.}

\item{bar_color}{A character string representing the color of the error bars (default is "#0085CA").}

\item{error_bars}{A logical value indicating whether to include error bars (default is TRUE).}
}
\value{
A ggplot object representing the generated subarea-specific biomass plots.
}
\description{
This function generates subarea-specific biomass plots based on indicator data.
}
\examples{
\dontrun{
# Example Usage:
make_subarea_plot(indicator_data = AI_INDICATOR,
                  indicator_name = "misc_species",
                  bar_color = "#0085CA",
                  error_bars = TRUE)
}

}
