% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBS_INDICATOR}
\alias{NBS_INDICATOR}
\title{Northern Bering Sea Indicator}
\format{
An object of class \code{list} of length 5.
}
\usage{
NBS_INDICATOR
}
\description{
ESR Indicator data for the Northern Bering Sea
}
\keyword{datasets}
