% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue_maps.R
\name{plot_cpue_maps}
\alias{plot_cpue_maps}
\title{Make CPUE maps}
\usage{
plot_cpue_maps(gapindex_cpue, crs = "EPSG:3338", breaks = NULL, fig_res = 300)
}
\arguments{
\item{gapindex_cpue}{A data.frame output by gapindex::calc_cpue()}

\item{crs}{Coordinate reference system for the map as a type recognized by sf::st_crs()}

\item{breaks}{Optional break values for CPUE plots. Uses package defaults if not provided.}

\item{fig_res}{Figure resolution for .png output files.}
}
\description{
Make CPUE maps
}
