% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_db.R
\name{plot_region_db}
\alias{plot_region_db}
\title{Create design-based index indicator plots for a region}
\usage{
plot_region_db(
  indicator_data,
  indicator_name,
  error_bars = TRUE,
  bar_color = "#0085CA"
)
}
\arguments{
\item{indicator_data}{A list containing indicator data, e.g. esrindex::EBS_INDICATOR, esrindex::GOA_INDICATOR.}

\item{indicator_name}{A character string specifying the indicator name included in esrindex::chapter_settings (e.g. "misc_species", "jellyfish").}

\item{error_bars}{A logical value indicating whether to include error bars (default is TRUE).}

\item{bar_color}{Color for error bars"#0085CA").}
}
\value{
A ggplot plot.
}
\description{
This function generates a region-specific biomass plot based on indicator data and saves the plot as a png.
}
\examples{
\dontrun{
plot_region_db(indicator_data = AI_INDICATOR,
                 indicator_name = "misc_species",
                 bar_color = "#0085CA",
                 error_bars = TRUE)
}
}
