% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_rema.R
\name{plot_region_rema}
\alias{plot_region_rema}
\title{Make ESR region plots with rema output}
\usage{
plot_region_rema(
  x,
  error_bar = TRUE,
  benchmarks = "none",
  point_color = "#0085CA",
  timeseries_color = "#000000",
  errorbar_color = "#000000",
  ribbon_fill = "grey50",
  hline_color = "grey50",
  append_filename = "",
  y_axis_min = NULL,
  set_unit = "kt"
)
}
\arguments{
\item{x}{Indicator list (e.g. esrindex::AI_INDICATOR)}

\item{error_bar}{Should annual observations include error bars?}

\item{benchmarks}{"none", "zscore", or "quantile"}

\item{point_color}{Color for points.}

\item{timeseries_color}{Color for time series line.}

\item{errorbar_color}{Color for error bar.}

\item{ribbon_fill}{Color for time series ribbon.}

\item{hline_color}{Color for Z-score/quantile lines.}

\item{append_filename}{Character vector to include in file name.}

\item{y_axis_min}{Optional. Minimum value for the y-axis.}

\item{set_unit}{Set units to use for output ("kt" = kiloton, "mt" = metric tons)}
}
\value{
The function doesn't return any value; it generates and saves plots based on the input data.
}
\description{
This function writes full region ESR abundance indicators for rema fits and writes them to ./plots/[region]/[indicator]_full_region.png
}
\examples{
\dontrun{
# Example:
                  error_bar = TRUE, 
                  benchmarks = "zscore")
}
}
