% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subarea_rema.R
\name{plot_subarea_rema}
\alias{plot_subarea_rema}
\title{Make ESR subarea/stratum plots with rema output}
\usage{
plot_subarea_rema(
  x,
  error_bar = TRUE,
  point_color = "#0085CA",
  timeseries_color = "#000000",
  errorbar_color = "#000000",
  color_palette = "noaa-blue",
  ribbon_fill = "grey50",
  hline_color = "grey50",
  benchmarks = "none",
  append_filename = "",
  set_unit = "kt"
)
}
\arguments{
\item{x}{Indicator list (e.g. esrindex::AI_INDICATOR)}

\item{error_bar}{Should annual observations include error bars?}

\item{point_color}{Color for points.}

\item{timeseries_color}{Color for time series line.}

\item{errorbar_color}{Color for error bar.}

\item{color_palette}{Color palette for bars (either "noaa-blue" or "brown-green").}

\item{ribbon_fill}{Color for time series ribbon.}

\item{hline_color}{Color for Z-score/quantile lines.}

\item{benchmarks}{"none", "zscore", or "quantile"}

\item{append_filename}{Character vector to include in file name.}

\item{set_unit}{Set units to use for output ("kt" = kiloton, "mt" = metric tons)}
}
\value{
The function doesn't return any value; it generates and saves plots based on the input data.
}
\description{
This function writes full region ESR abundance indicators for rema fits and writes them to ./plots/[region]/[indicator]_subarea_bar.png and ./plots/[region]/[indicator]_subarea_point.png
}
\examples{
\dontrun{
# Example:
                   error_bar = TRUE, 
                   benchmarks = "zscore")
}

}
