% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_biomass.R
\name{plot_top_biomass}
\alias{plot_top_biomass}
\title{Biomass index trends for AFSC bottom trawl survey regions}
\usage{
plot_top_biomass(
  channel,
  region,
  max_year,
  n_species = NULL,
  species_codes = NULL,
  legend_position = "top-left"
)
}
\arguments{
\item{channel}{An active ODBC connection channel object generated via `RODBC`.}

\item{region}{Character string specifying the survey region. Options are:
 * `"ebs"`: Eastern Bering Sea (area_id = 99901, survey_definition_id = 98)
 * `"goa"`: Gulf of Alaska (area_id = 99903, survey_definition_id = 47)
 * `"ai"`: Aleutian Islands (area_id = 99904, survey_definition_id = 52)
* `"nbs"`: Northern Bering Sea (area_id = 99905, survey_definition_id = 143)}

\item{max_year}{Numeric or integer specifying the final survey year to query and base species biomass ranking upon.}

\item{n_species}{Integer. Number of top biomass species to select. Default is `NULL`. Ignored if `species_codes` is provided.}

\item{species_codes}{Vector of numeric RACE species codes to use. Overrides `n_species`. Default is `NULL`.}

\item{legend_position}{Character string setting the relative position of the plot legend inside the panel. Options are `"top-left"` (default) or `"top-right"`.}
}
\value{
A named `list` containing:
\describe{
  \item{biomass_plot}{A `ggplot` object with design-based biomass index plots.}
  \item{biomass_data}{A `data.frame` containing the raw query records from GAP_PRODUCTS.AKFIN_BIOMASS.}
}
}
\description{
Queries GAP_PRODUCTS AKFIN biomass tables and generates a  `ggplot2` time-series plot with mean 
design-based biomass index and 95% CI. Exports PNG image to the /plots/<region>/ directory.
}
\details{
The function writes two 6x6 inch PNG files to /plots/<region>/:
* `<region>_db_top_species.png` (with main title)
* `<region>_db_top_species_no_title.png` (without title)
}
