% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_esr_xml.R
\name{read_esr_xml}
\alias{read_esr_xml}
\title{Read tag data from an ESR XML template file}
\usage{
read_esr_xml(xml_path)
}
\arguments{
\item{xml_path}{The path to the ESR XML template file.}
}
\value{
A list containing information extracted from XML tags.
}
\description{
This function reads tag data from an ESR (Ecosystem Status Report) XML template file.
}
