% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_settings}
\alias{region_settings}
\title{ESR region settings}
\format{
A list with settings for each region, where each region has 3 variables:
  \describe{
      \item{esr_subarea_id}{Corresponds with subarea AREA_ID from gapindex.}
      \item{esr_area_id}{Corresponds with region-wide AREA_ID from gapindex.}
      \item{min_year}{Starting year}
    }
}
\usage{
region_settings
}
\description{
This dataset contains subarea and area IDs, and the minimum years to use for biomass.
}
\keyword{datasets}
