% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rema.R
\name{run_rema}
\alias{run_rema}
\title{Fit random effects model to biomass index time series}
\usage{
run_rema(x, region, group_name, zero_assumption)
}
\arguments{
\item{x}{A data.frame containing SURVEY, YEAR, BIOMASS_MT, and CV columns}

\item{region}{A character vector indicating the region  ("AI", "GOA", "EBS", or "NBS")}

\item{group_name}{Character vector indicating the name of the group (e.g. "Sponges")}

\item{zero_assumption}{Assumption for zero biomass observations ("tweedie", "na", "small_constant")}
}
\description{
Fit a random effects model using the rema package.
}
