% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_stratum_order}
\alias{set_stratum_order}
\title{Set stratum plotting order}
\usage{
set_stratum_order(stratum, region)
}
\arguments{
\item{stratum}{Stratum as a character vector}

\item{region}{Character vector of length one indicating whether the region is AI, GOA, or BS}
}
\description{
Set stratum plotting order
}
