% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_groups}
\alias{species_groups}
\title{Species Groups}
\format{
A data frame with 3 variables:
  \describe{
    \item{group_name}{Character vector, names of taxonomic groups.}
    \item{min_code}{Numeric vector, minimum SPECIES_CODE for each group.}
    \item{max_code}{Numeric vector, maximum SPECIES_CODE for each group.}
    \item{complex}{Logical indicating whether the group is a multi-species complex.}
  }
}
\usage{
species_groups
}
\description{
This dataset contains group names and RACEBASE SPECIES_CODE ranges for species groups used in ESRs.
}
\keyword{datasets}
