% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_groups_ns}
\alias{species_groups_ns}
\title{Non-sequential species groups}
\format{
An object of class \code{list} of length 3.
}
\usage{
species_groups_ns
}
\description{
This named list contains species codes for species categories that are non-sequential.
}
\keyword{datasets}
