% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_figure.R
\name{add_figure}
\alias{add_figure}
\title{Add Figure to QMD or RMD document}
\usage{
add_figure(x)
}
\arguments{
\item{x}{A list from esrindex::read_esr_xml containing figure number (fig_num), caption (fig_caption), and figure path (fig_path).}
}
\description{
This function uses outputs from esrindex::read_esr_xml to add figures to an RMD or QMD template document.
}
\examples{
# Example usage:
fig_info <- list(fig_number = 1, caption = "Example Figure", fig_path = "example_plot.png")
quarto_code <- add_figure(fig_info)

}
