% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esrindex-package.R
\docType{package}
\name{esrindex-package}
\alias{esrindex}
\alias{esrindex-package}
\title{esrindex: Make index products for Ecosystem Status Reports}
\description{
Make abundance index products for Ecosystem Status Reports.
}
\keyword{internal}
