% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{esrindex_pal}
\alias{esrindex_pal}
\title{Color palette for esrindex}
\usage{
esrindex_pal(...)
}
\arguments{
\item{...}{Optional arguments passed to colorRampPalette}
}
\description{
This is the 'oceans' color palette from nmfspalette
}
