% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_data.R
\name{get_group_data}
\alias{get_group_data}
\title{Get index data for ESR species and groups}
\usage{
get_group_data(
  region,
  channel = NULL,
  zero_assumption = "na",
  rema_by_stratum = TRUE
)
}
\arguments{
\item{region}{A character string specifying the region for which data should be retrieved ("GOA", "AI", "EBS", or "NBS")}

\item{channel}{An optional parameter specifying an RODBC database channel. If not provided, a connection is established.}

\item{zero_assumption}{Assumption for zero biomass observations ("tweedie", "na", "small_constant"). Default = "na"}

\item{rema_by_stratum}{Logical indicating whether to estimate biomass by survey stratum (TRUE) or estimate biomass from the total biomass index (FALSE).}
}
\value{
A list containing two data frames: timeseries and mean_sd.
\itemize{
  \item \code{timeseries}: Data frame with biomass time series for each subarea and species group.
  \item \code{mean_sd}: Summary data frame with mean and standard deviation of biomass for each subarea and species group.
}
}
\description{
Calculates group-level biomass index for a specified region.
}
\details{
This function retrieves invertebrate data for the specified region using the provided channel.
If the channel is not provided, the function uses the default connected channel obtained from \code{esrindex::get_connected()}.
}
\examples{
\dontrun{
# Example Usage:
data <- get_group_data(region = "GOA")
}
}
