% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subarea_rema.R
\name{plot_subarea_rema}
\alias{plot_subarea_rema}
\title{Make ESR subarea/stratum plots with rema output}
\usage{
plot_subarea_rema(
  x,
  error_bar = TRUE,
  benchmarks = "none",
  append_filename = ""
)
}
\arguments{
\item{x}{Indicator list (e.g. esrindex::AI_INDICATOR)}

\item{error_bar}{Should annual observations include error bars?}

\item{benchmarks}{"none", "zscore", or "quantile"}

\item{append_filename}{Character vector to include in file name.}
}
\value{
The function doesn't return any value; it generates and saves plots based on the input data.
}
\description{
This function writes full region ESR abundance indicators for rema fits and writes them to ./plots/[region]/[indicator]_subarea_bar.png and ./plots/[region]/[indicator]_subarea_point.png
}
\examples{
\dontrun{
# Example:
                   error_bar = TRUE, 
                   benchmarks = "zscore")
}

}
